spark-analytics/requirements.txt
pyspark==3.5.0
pymongo==4.6.1
python-dotenv==1.0.0
pandas>=2.1.0
requests==2.31.0
